+++ crates/tusker-query/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support checked PostgreSQL array parameter and row types
+- Add `#[derive(QueryComposite)]` for structurally checked PostgreSQL composite parameter and row types
+
 ## [0.2.1] - 2026-05-23
 
 ### Changed